✅(backend) extract demo data generation
Extract demo data generation to use it in order export test.
